Added service layer to API
