+++ Backend/FitnessTracker.WebAPI/Controllers/AerobicTrainingsController.cs
@@ -9,6 +9,7 @@ using FitnessTracker.WebAPI.ApiResponse;
 using Microsoft.AspNetCore.Http;
 using System;
 using System.Text.Json;
+using FitnessTracker.WebAPI.Interfaces;
 
 namespace FitnessTracker.WebAPI.Controllers
 {
@@ -20,250 +21,62 @@ namespace FitnessTracker.WebAPI.Controllers
         private readonly IHttpContextAccessor _http;
         private readonly string ExportContentType = "application/json";
         private readonly string DefaultExportFileName = "Aerobic Trainings.json";
+        private readonly IAerobicTrainingsService _service;
 
-        public AerobicTrainingsController(TrainingsContext context, IHttpContextAccessor httpContextAccessor)
+        public AerobicTrainingsController(TrainingsContext context, IHttpContextAccessor httpContextAccessor, IAerobicTrainingsService service)
         {
             _context = context;
             _http = httpContextAccessor; 
+            _service = service;
         }
 
         [HttpGet]
         public async Task<ApiResponse<IEnumerable<AerobicTrainingDto>>> GetAerobicTrainings()
         {
-          if (_context.AerobicTrainings == null)
-          {
-                return new ApiResponse<IEnumerable<AerobicTrainingDto>>
-                {
-                    IsSuccess = false,
-                    StatusCode = 404,
-                    ErrorMessage = "No Aerobic Trainings Data Found",                   
-                };
-          }
+            var response = await _service.GetAerobicTrainings();
 
-            var aerobicTrainings = await _context.AerobicTrainings.Select(at => new AerobicTrainingDto
-            {
-                ActivityDurationMinutes = at.ActivityDurationMinutes,
-                ActivityDate = at.ActivityDate,
-                ActivityType = at.ActivityType,
-                CalorieBurnt = at.CalorieBurnt,
-                AerobicTrainingId = at.AerobicTrainingId,
-            }).ToListAsync();
-
-            return new ApiResponse<IEnumerable<AerobicTrainingDto>>
-            {
-                IsSuccess = true,
-                StatusCode = 200,
-                Data = aerobicTrainings,
-            };
+            return response;
         }
 
         [HttpGet("user/{username}")]
         public async Task<ApiResponse<IEnumerable<AerobicTrainingDto>>> GetAerobicTrainingsForUser(string username)
         {
-            if (_context.AerobicTrainings == null)
-            {
-                return new ApiResponse<IEnumerable<AerobicTrainingDto>>
-                {
-                    IsSuccess = false,
-                    StatusCode = 404,
-                    ErrorMessage = "No Aerobic Trainings Data Found for this user",
-                };
-            }
+            var response = await _service.GetAerobicTrainingsForUser(username);
 
-            var userId = _context.Users?.FirstOrDefault(u => u.UserName == username)?.UserId;
-            var aerobicTrainings = await _context.AerobicTrainings.Where(at => at.UserId == userId).Select(at => new AerobicTrainingDto
-            {
-                ActivityDurationMinutes = at.ActivityDurationMinutes,
-                ActivityDate = at.ActivityDate,
-                ActivityType = at.ActivityType,
-                CalorieBurnt = at.CalorieBurnt,
-                AerobicTrainingId = at.AerobicTrainingId,
-            }).ToListAsync();
-
-            return new ApiResponse<IEnumerable<AerobicTrainingDto>>
-            {
-                IsSuccess = true,
-                StatusCode = 200,
-                Data = aerobicTrainings,
-            };
+            return response;
         }
 
         [HttpGet("{id}")]
         public async Task<ApiResponse<AerobicTrainingDto>> GetAerobicTraining(Guid id)
         {
-          if (_context.AerobicTrainings == null)
-          {
-                return new ApiResponse<AerobicTrainingDto>
-                {
-                    IsSuccess = false,
-                    StatusCode = 404,
-                    ErrorMessage = "Aerobic Trainings Not Defined",
-                };
-            }
-            var aerobicTraining = await _context.AerobicTrainings.FindAsync(id);
+            var response = await _service.GetAerobicTraining(id);
 
-            if (aerobicTraining == null)
-            {
-                return new ApiResponse<AerobicTrainingDto>
-                {
-                    IsSuccess = false,
-                    StatusCode = 404,
-                    ErrorMessage = "Aerobic Training Not Found",
-                };
-            }
-
-            return new ApiResponse<AerobicTrainingDto>
-            {
-                IsSuccess = true,
-                StatusCode = 200,
-                Data = new AerobicTrainingDto
-                {
-                    ActivityDurationMinutes = aerobicTraining.ActivityDurationMinutes,
-                    ActivityDate = aerobicTraining.ActivityDate,
-                    ActivityType = aerobicTraining.ActivityType,
-                    CalorieBurnt = aerobicTraining.CalorieBurnt,
-                    AerobicTrainingId = aerobicTraining.AerobicTrainingId,
-                }
-            };
+            return response;
         }
 
         [HttpPut("{id}")]
         public async Task<ApiResponse<AerobicTrainingDto>> PutAerobicTraining(Guid id, AerobicTrainingDto aerobicTraining)
         {
-            var editTraining = _context.AerobicTrainings.Find(id);
+            var response = await _service.PutAerobicTraining(id, aerobicTraining);
 
-            if (editTraining == null)
-            {
-                return new ApiResponse<AerobicTrainingDto>
-                {
-                    IsSuccess = false,
-                    StatusCode = 404,
-                    ErrorMessage = "Training Not Found",
-                };
-            }
-
-            try
-            {
-                editTraining.ActivityDurationMinutes = aerobicTraining.ActivityDurationMinutes;
-                editTraining.CalorieBurnt = aerobicTraining.CalorieBurnt;
-                editTraining.ActivityType = aerobicTraining.ActivityType;
-                editTraining.ActivityDate = aerobicTraining.ActivityDate;
-                editTraining.ModificationTime = DateTime.UtcNow;
-
-                await _context.SaveChangesAsync();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!AerobicTrainingExists(id))
-                {
-                    return new ApiResponse<AerobicTrainingDto>
-                    {
-                        IsSuccess = false,
-                        StatusCode = 404,
-                        ErrorMessage = "Training Not Found",
-                    };
-                }
-                else
-                {
-                    throw;
-                }
-            }
-
-            return new ApiResponse<AerobicTrainingDto>
-            {
-                IsSuccess = true,
-                StatusCode = 204,
-                Data = new AerobicTrainingDto 
-                { 
-                    ActivityDate = aerobicTraining.ActivityDate,
-                    ActivityDurationMinutes = aerobicTraining.ActivityDurationMinutes,
-                    CalorieBurnt = aerobicTraining.CalorieBurnt,
-                    ActivityType = aerobicTraining.ActivityType,
-
-                }
-            };
+            return response;
         }
 
         [HttpPost]
         [Authorize]
         public async Task<ApiResponse<AerobicTrainingDto>> PostAerobicTraining(AerobicTrainingDto aerobicTraining)
         {
-          if (_context.AerobicTrainings == null)
-          {
-                return new ApiResponse<AerobicTrainingDto>
-                {
-                    IsSuccess = false,
-                    StatusCode = 404,
-                    ErrorMessage = "Aerobic Trainings Not Defined",
-                };
-            }
+            var response = await _service.PostAerobicTraining(aerobicTraining);
 
-
-            var username = _http.HttpContext?.User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
-            var userId = _context.Users.First(u => u.UserName == username).UserId;
-
-            if (userId == Guid.Empty) 
-            {
-                return new ApiResponse<AerobicTrainingDto>
-                {
-                    IsSuccess = false,
-                    StatusCode = 404,
-                    ErrorMessage = "User Not Found",
-                };
-            }
-
-            var newAerobicTraining = new AerobicTraining(userId, aerobicTraining.ActivityType!, aerobicTraining.ActivityDurationMinutes, aerobicTraining.CalorieBurnt, aerobicTraining.ActivityDate);
-
-            _context.AerobicTrainings.Add(newAerobicTraining);
-            await _context.SaveChangesAsync();
-
-
-            return new ApiResponse<AerobicTrainingDto>
-            {
-                IsSuccess = true,
-                StatusCode = 200,
-                Data = new AerobicTrainingDto
-                {
-                    ActivityDurationMinutes = aerobicTraining.ActivityDurationMinutes,
-                    ActivityType = aerobicTraining.ActivityType,
-                    CalorieBurnt = aerobicTraining.CalorieBurnt,
-                    ActivityDate = aerobicTraining.ActivityDate
-                }
-            };
+            return response;
         }
 
         [HttpDelete("{id}")]
-        public async Task<ApiResponse<string>> DeleteAerobicTraining(Guid id)
+        public async Task<ApiResponse<AerobicTrainingDto>> DeleteAerobicTraining(Guid id)
         {
-            if (_context.AerobicTrainings == null)
-            {
-                return new ApiResponse<string>
-                {
-                    IsSuccess = false,
-                    StatusCode = 404,
-                    ErrorMessage = "Aerobic Trainings Not Defined",
-                };
-            }
-            var aerobicTraining = await _context.AerobicTrainings.FindAsync(id);
-            if (aerobicTraining == null)
-            {
-                return new ApiResponse<string>
-                {
-                    IsSuccess = false,
-                    StatusCode = 404,
-                    ErrorMessage = "Aerobic Training Not Found",
-                };
-            }
+            var response = await _service.DeleteAerobicTraining(id);
 
-            _context.AerobicTrainings.Remove(aerobicTraining);
-            await _context.SaveChangesAsync();
-
-            return new ApiResponse<string>
-            {
-                IsSuccess = false,
-                StatusCode = 204,
-                Data = "No Content"
-            };
+            return response;
         }
         [HttpGet("export")]
         public async Task<IActionResult> exportTrainings()
@@ -308,13 +121,8 @@ namespace FitnessTracker.WebAPI.Controllers
             var exportData = JsonSerializer.Serialize(trainingsData, options);
             var fileBytes = System.Text.Encoding.UTF8.GetBytes(exportData);
 
-            Response.Headers.Add("Content-Disposition", "attachment; filename=Aerobic Trainings.json");
+            Response.Headers.Append("Content-Disposition", "attachment; filename=Aerobic Trainings.json");
             return File(fileBytes, ExportContentType, DefaultExportFileName);
-        }
-
-        private bool AerobicTrainingExists(Guid id)
-        {
-            return (_context.AerobicTrainings?.Any(e => e.AerobicTrainingId == id)).GetValueOrDefault();
         }
     }
 }
